add new controller - service
#85
