+++ src/main/java/es/ucm/fdi/iw/SecurityConfig.java
@@ -65,7 +65,8 @@ public class SecurityConfig {
 						.permitAll() // <-- public access
 						.requestMatchers("/admin/**")
 						.hasRole("ADMIN") // <-- administration
-						.requestMatchers("/user/**", "/perfil", "/amigos", "/sala-espera", "/configuracion-partida", "/partida", "/resultados")
+						.requestMatchers("/user/**", "/perfil", "/amigos", "/sala-espera", "/configuracion-partida",
+								"/partida", "/resultados")
 						.hasRole("USER") // <-- logged-in users
 						.anyRequest().authenticated())
 				.formLogin(formLogin -> formLogin
